pyproject.toml
[tool.poetry]
name = "liaksej-tiny-cloud-backend"
version = "1.0.0-alpha"
description = ""
authors = ["Liaksej <tarusov.liaksej@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
python-dotenv = "^1.0.0"
psycopg = {extras = ["binary"], version = "^3.1.13"}
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.4"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.0"}
django-allauth = "^0.58.2"
dj-rest-auth = "^5.0.2"
django-cors-headers = "^4.3.1"
django-debug-toolbar = "^4.2.0"
black = "^23.11.0"
pytest = "^7.4.3"
mixer = "^7.2.2"
pytest-django = "^4.7.0"
pytest-cov = "^4.1.0"
beautifulsoup4 = "^4.12.2"
pytest-timeout = "^2.2.0"
pytest-mock = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test = "pytest ."
test_coverage = "pytest --cov-. --cov-report=html"
